Store data in directory nodes

+++ include/spaix/draw_dot.hpp
@@ -71,7 +71,7 @@ draw_dat_dot(std::ostream& os,
 
 template <class Tree>
 void
-draw_dot(std::ostream& os, const Tree& tree, const unsigned max_depth = 0)
+draw_dot(std::ostream& os, const Tree& tree, const size_t max_depth = 0u)
 {
   using Key    = typename Tree::Key;
   using Data   = typename Tree::Data;
